+++ wiseasily/src/main/java/wiseasily/util/ScanFilter.java
@@ -3,9 +3,7 @@ package wiseasily.util;
 import android.net.wifi.ScanResult;
 import android.support.annotation.Nullable;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import wiseasily.util.util.ProximityUtils;
@@ -23,7 +21,9 @@ public class ScanFilter {
     @Nullable
     private String ssid;
     @Nullable
-    private Set<String> ssids;
+    private Set<String> includeSsids;
+    @Nullable
+    private Set<String> excludeSsids;
     @Nullable
     private Set<Integer> channels;
     @Nullable
@@ -34,10 +34,11 @@ public class ScanFilter {
      *
      * @param mac      Ethernet MAC address, e.g., XX:XX:XX:XX:XX:XX where each X is a hex digit
      * @param ssid     service set identifier of the 802.11 network
-     * @param channels Set of channel numbers
-     * @throws IllegalArgumentException if {@code mac}'s length > 17 chars or {@code ssid}'s length > 32 chars
+     * @param excludeSsids SSIDS that want to show in scan
+     * @param excludeSsids SSIDS that wont to show in scan
+     *@param channels Set of channel numbers  @throws IllegalArgumentException if {@code mac}'s length > 17 chars or {@code ssid}'s length > 32 chars
      */
-    public ScanFilter(@Nullable String mac, @Nullable String ssid, @Nullable Set<String> ssids, @Nullable Set<Integer> channels, @Nullable Proximity proximity) {
+    public ScanFilter(@Nullable String mac, @Nullable String ssid, @Nullable Set<String> includeSsids, Set<String> excludeSsids, @Nullable Set<Integer> channels, @Nullable Proximity proximity) {
         if (mac != null && mac.length() > 17) {
             throw new IllegalArgumentException("Mac longer than 17 chars");
         }
@@ -53,18 +54,26 @@ public class ScanFilter {
         }
         this.channels = channels;
 
-        if(ssids!=null){
+        if(includeSsids !=null){
             Set<String> ssidsAssign = new HashSet<>();
-            for(String id : ssids){
+            for(String id : includeSsids){
                 ssidsAssign.add(id.toLowerCase());
             }
-            this.ssids = ssidsAssign;
+            this.includeSsids = ssidsAssign;
+        }
+
+        if(this.excludeSsids !=null){
+            Set<String> ssidsExclude = new HashSet<>();
+            for(String id : excludeSsids){
+                ssidsExclude.add(id.toLowerCase());
+            }
+            this.excludeSsids = ssidsExclude;
         }
         this.proximity = proximity;
     }
 
     public boolean matchesStart(ScanResult scanResult) {
-        return scanResult != null && (mac == null || scanResult.BSSID.toLowerCase().startsWith(mac)) && (ssid == null || scanResult.SSID.toLowerCase().startsWith(ssid)) && (ssids == null || ssids.isEmpty() || ssids.contains(scanResult.SSID.toLowerCase())) && (channels == null || channels.contains(WifiUtils.toChannel(scanResult.frequency))) && (proximity == null || proximity == ProximityUtils.getProximity(scanResult.level, scanResult.frequency));
+        return scanResult != null && (mac == null || scanResult.BSSID.toLowerCase().startsWith(mac)) && (ssid == null || scanResult.SSID.toLowerCase().startsWith(ssid)) && (includeSsids == null || includeSsids.isEmpty() || includeSsids.contains(scanResult.SSID.toLowerCase())) && (excludeSsids == null || excludeSsids.isEmpty() || excludeSsids.contains(scanResult.SSID.toLowerCase())) && (channels == null || channels.isEmpty() || channels.contains(WifiUtils.toChannel(scanResult.frequency))) && (proximity == null || proximity == ProximityUtils.getProximity(scanResult.level, scanResult.frequency));
 
     }
 }