+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ##### Bug Fix
-* Handle if ssids empty
+* Handle if channel empty
+* Handle if excludeSsids
+
+
+#### 0.1.5-beta (2018-01-11)
+##### Bug Fix
+* Handle if includeSsids empty
 
 
 #### 0.1.4-beta (2018-01-11)